Remove variants file
